DSS-18337: Remove Algorithm specific methods from Interfaces & HBT for Pig
