+++ app/com/linkedin/drelephant/tuning/hbt/SparkHBTParamRecommender.java
@@ -1,4 +1,4 @@
-package com.linkedin.drelephant.tuning.engine;
+package com.linkedin.drelephant.tuning.hbt;
 
 import java.util.HashMap;
 import java.util.Map;
